Refactoring file names to camelCase
